feat: 모델 추가

+++ LongLife/src/main/java/com/oraclejava/longlife/model/Exercise.java
@@ -12,7 +12,7 @@ public class Exercise {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "exercise_id")
-    private int exerciseId;
+    private Long exerciseId;
 
     private String type1;
